feat(metrics): shared process-artifact filter + shipped-only diff_loc, drop commits column — minimal green (Story h8)
harness/lib/process-artifacts.ts is the new single-source filter (anticipates
h10): PROCESS_ARTIFACT_PREFIXES excludes docs/plans/, docs/retrospectives/,
docs/status.d/. diffLocForSha in loop-metrics.ts now delegates its numstat
parsing to sumShippedDiffLoc. Drop the dead `commits` column/field
(always 1 post-squash) from LoopRow, BuildLoopRowInput, and CSV_COLUMNS;
delete countStoryCommits.

weight_ratio keeps its plan_loc/diff_loc formula but its meaning flips now
that diff_loc is shipped-only — comment added at the computation site.

Regenerated docs/metrics/loop.csv: 5-column header, maint-18's weight_ratio
flips from 0.64 (reported, ceremony-inflated) to 26 (shipped-only) —
confirms the metric-fix direction ahead of slice 5's retro table.
