+++ docs/metrics/loop.csv
@@ -1,41 +1,42 @@
-story_id,plan_loc,diff_loc,commits,weight_ratio,retro_loop_metrics
-2.2,252,1155,1,0.21818181818181817,true
-2.3,329,1238,1,0.265751211631664,true
-2.4,321,2142,1,0.14985994397759103,true
-2.5,608,352,2,1.7272727272727273,true
-3.1,390,1288,1,0.3027950310559006,true
-3.2,193,1470,1,0.13129251700680272,true
-3.3,237,1985,1,0.11939546599496222,true
-3.4,279,2127,1,0.1311706629055007,true
-3.5,313,2765,1,0.1132007233273056,true
-A,208,666,1,0.3123123123123123,false
-B,298,1070,1,0.27850467289719627,false
-C,404,2083,1,0.19395103216514642,false
-D,313,1789,1,0.17495807713806596,false
-ddd-1,163,707,1,0.23055162659123055,false
-h1,363,1231,1,0.29488220958570266,false
-h2,214,327,1,0.654434250764526,false
-h3,383,535,1,0.7158878504672898,true
-h4,265,1864,1,0.14216738197424894,true
-h5,295,445,1,0.6629213483146067,true
-h6,361,2594,1,0.13916730917501927,true
-h7,234,615,1,0.3804878048780488,true
-maint-01,276,395,1,0.6987341772151898,true
-maint-02,109,182,1,0.5989010989010989,true
-maint-03,173,408,1,0.42401960784313725,true
-maint-04,197,538,1,0.36617100371747213,true
-maint-05,217,1288,1,0.16847826086956522,true
-maint-06,219,664,1,0.32981927710843373,true
-maint-07,164,246,1,0.6666666666666666,true
-maint-08,178,359,1,0.4958217270194986,true
-maint-09,172,352,1,0.48863636363636365,false
-maint-10,275,1516,1,0.18139841688654354,false
-maint-11,291,1025,1,0.2839024390243902,false
-maint-12,269,518,1,0.5193050193050193,false
-maint-13,182,371,1,0.49056603773584906,false
-maint-14,158,364,1,0.4340659340659341,false
-maint-15,117,173,1,0.6763005780346821,false
-maint-16,218,614,1,0.3550488599348534,false
-maint-17,165,n/a,0,n/a,false
-maint-18,104,163,1,0.6380368098159509,false
-maint-19,128,353,1,0.3626062322946176,false
+story_id,plan_loc,diff_loc,weight_ratio,retro_loop_metrics
+2.2,252,853,0.2954279015240328,true
+2.3,329,862,0.38167053364269143,true
+2.4,321,1766,0.18176670441676104,true
+2.5,608,157,3.872611464968153,true
+3.1,390,848,0.45990566037735847,true
+3.2,193,1221,0.15806715806715807,true
+3.3,237,1677,0.1413237924865832,true
+3.4,279,1767,0.15789473684210525,true
+3.5,313,2367,0.13223489649345163,true
+A,208,411,0.5060827250608273,false
+B,298,718,0.415041782729805,false
+C,404,1609,0.25108763206960844,false
+D,313,1400,0.22357142857142856,false
+ddd-1,163,520,0.31346153846153846,false
+h1,363,795,0.45660377358490567,false
+h2,214,49,4.36734693877551,false
+h3,383,80,4.7875,true
+h4,265,1518,0.17457180500658762,true
+h5,295,59,5,true
+h6,361,2128,0.16964285714285715,true
+h7,234,285,0.8210526315789474,true
+h8,162,n/a,n/a,false
+maint-01,276,59,4.677966101694915,true
+maint-02,109,20,5.45,true
+maint-03,173,173,1,true
+maint-04,197,283,0.696113074204947,true
+maint-05,217,1011,0.21463897131552917,true
+maint-06,219,387,0.5658914728682171,true
+maint-07,164,21,7.809523809523809,true
+maint-08,178,118,1.5084745762711864,true
+maint-09,172,157,1.0955414012738853,false
+maint-10,275,1204,0.22840531561461794,false
+maint-11,291,712,0.40870786516853935,false
+maint-12,269,193,1.3937823834196892,false
+maint-13,182,150,1.2133333333333334,false
+maint-14,158,167,0.9461077844311377,false
+maint-15,117,18,6.5,false
+maint-16,218,71,3.0704225352112675,false
+maint-17,165,n/a,n/a,false
+maint-18,104,4,26,false
+maint-19,128,171,0.7485380116959064,false
